requirements.txt
# Usage: pip install -r requirements.txt

# Base
matplotlib>=3.7.1
numpy>=1.24.2
opencv-python>=4.7.0.72
Pillow>=9.4.0
torch>=1.13.1
torchvision>=0.13.1
tqdm>=4.65.0

# Extra
pre-commit
flake8
ufmt
pydocstyle
